pyproject.toml
[tool.poetry]
name = "wap"
version = "0.1.0"
description = "Build and upload your WoW addons"
authors = ["Tim Martin <tim@timmart.in>"]
license = "GPLv3"

[tool.poetry.dependencies]
python = "^3.9"
attrs = "^20.3.0"
requests = "^2.25.1"
click = "^7.1.2"
strictyaml = "^1.3.2"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
mypy = "^0.800"
pytest = "^6.2.2"
flake8 = "^3.8.4"
pyfakefs = "^4.3.3"
pytest-cov = "^2.11.1"
codecov = "^2.1.11"
requests-mock = "^1.8.0"
pytest-mock = "^3.5.1"
isort = "^5.7.0"
bump2version = "^1.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
wap = 'wap.__main__:main'
